Add files via upload
This is the code for Ronroy Chatbot 

This project integrates Ollama's Llama 3.2 model with a Gradio interface to create "Ronroy," a lively 19-year-old AI companion known for his sharp wit and playful attitude. Users can engage in dynamic and entertaining conversations with Ronroy, who brings humor, sarcasm, and warmth to every interaction.